Transfer theme surface system to theme dir, export
