+++ src/theme/index.ts
@@ -1,3 +1,4 @@
 export { ThreadTheme } from './thread-theme';
 export { ThreadScript } from './thread-script';
 export { ThemeProvider, useThemeMode } from './theme-provider';
+export { SurfaceLevelMap } from './theme-surface-system';
